Store the Portal name instead of a pointer in PL/Python cursors.
The pointer to the Portal might become invalid if the subtransaction
where the cursor has been created gets aborted and coupled with the
ability to create explicit subtransactions leads to a way to crash the
backend by continuing to use a PLyCursor object that's been created in
an aborted subxact.

Report and test cases from a review by Steve Singer.